refactor: complete tests

+++ database/migrations/2022_08_09_100000_create_media_variant_table.php
@@ -8,13 +8,16 @@ class CreateMediaVariantTable extends Migration
 {
     public function up(): void
     {
-        Schema::create($this->prefix.'media_product_variant', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('media_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_variant_id')->constrained($this->prefix.'product_variants')->onDelete('cascade');
-            $table->boolean('primary')->default(false)->index();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable($this->prefix.'media_product_variant')) {
+            Schema::create($this->prefix.'media_product_variant', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('media_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_variant_id')->constrained($this->prefix.'product_variants')->onDelete('cascade');
+                $table->boolean('primary')->default(false)->index();
+                $table->smallInteger('position')->default(1)->index();
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
